Added component inspector; Added zlib static lib; Added hierarchy panel and inspector panel; Improved release build by not needing assimp dll; Fixed various compiler warnings.

+++ engine/src/util/time.cpp
@@ -56,7 +56,7 @@ namespace emerald {
 	}
 
 	float Time::getTimeMs() {
-		return timeGetTime();
+		return (float)timeGetTime();
 
 	}
 	float Time::getTimeS() {
